Create week 02 Homework 1.1
This week 02 Homework 1.1 is based on the week 02 of past cohort week assessment. The purpose of this is to deepen and sharpen plutus core development. You are to re-construct the script by correcting everything that is wrong and explain to a simplel person line by line, word by word the purpose of the code. Take is as game of plutus puzzle to put together.